dashboard_service/base/tasks.py
from celery import shared_task
import requests
from .models import DailyBooking, MonthlyBooking
from datetime import datetime

DATA_PROVIDER_URL = 'http://data-provider-service-url/events/'

@shared_task
def sync_dashboard_data():
    # Example of fetching data from Data Provider's GET /events endpoint
    response = requests.get(DATA_PROVIDER_URL, params={'limit': 1000, 'offset': 0})
    if response.status_code == 200:
        events = response.json()
        # Process events and update DailyBooking and MonthlyBooking models accordingly
        for event in events:
            # Parse the event data and update the models
            print(event)
            pass
    else:
        # Handle error
        print(response.data)
        pass
